Install Jest for CI and add a test
