+++ .github/workflows/jest.yml
@@ -1,0 +1,21 @@
+name: Jest
+
+on: [push]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v1
+    - name: Use Node.js 10.x
+      uses: actions/setup-node@v1
+      with:
+        node-version: 10.x
+    - name: npm run test
+      run: |
+        npm install
+        npm run test
+      env:
+        CI: true
