+++ package.json
@@ -7,7 +7,8 @@
     "start": "next start",
     "create-sendgrid-secret": "now secret add tji-sendgrid-key YOUR_API_KEY_HERE",
     "lint": "eslint .",
-    "lint:fix": "eslint . --fix"
+    "lint:fix": "eslint . --fix",
+    "test": "jest"
   },
   "dependencies": {
     "@sendgrid/mail": "^6.3.1",
@@ -57,6 +58,7 @@
     "eslint-plugin-prettier": "^3.0.1",
     "eslint-plugin-react": "^7.12.4",
     "eslint-plugin-react-hooks": "^1.3.0",
+    "jest": "^24.9.0",
     "prettier": "^1.16.4"
   },
   "eslintConfig": {
